During join/leave have new vnodes act as proxies until handoff occurs
NOTE: This is an **experimental** patch written to solve a specific problem
for a specific user.  This code is solely the work of the author and
has not been blessed by Basho even though the author is an employee of
the company.  Use at your own risk, dragons may very well lie ahead.

Currently, when a node joins/leaves the cluster the ring is changed
and gossiped immediately, thus causing the preflist to change as well.
Depedning on how the partition claim went a great shuffle [of
partitions] may occur, causing some preflists to point to all new vnodes.
This is a problem since a new vnode starts with an empty state and
must wait for the old vnode to handoff it's data.  In this case,
not_founds will ensue until handoff has finished, regardless of the fact
that the data is there.

This patch augments the behavior of the new vnode to act as proxy
of the old one (what I refer to as the predecessor in the code)
until handoff has finished.  This way data is still available even
in the case where a great partition shuffle occurs and the preflist
points to all new vnodes.  The new vnode will cache it's
reults locally both avoiding extra hops when unnecessary and allowing
the predcessor to idle in order for handoff to start.

Furthermore, handoff will occur in an async fashion allowing the
predcessor to both handoff data **and** service proxied requests.